Bit of renaming in utils folder
