+++ assets/storage/options.json
@@ -11,6 +11,6 @@
         "warn": true,
         "info": true,
         "debug": true,
-        "trace": true
+        "trace": false
     }
 }
