+++ src/testbed.py
@@ -1,4 +1,4 @@
 from utils.storageHandler import storage
-from utils.consoleHandler import console
+from utils.console import console
 from utils.sceneHandler import scene
-from utils.timeHandler import date, clock, Timer, Chrono, delay
+from utils.timeToolbox import date, clock, Timer, Chrono, delay
